Reworked project to add testing framework.
