+++ Zelda.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 15.0.28307.271
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Zelda", "Zelda\Zelda.vcxproj", "{EBC6D2CA-29E7-4666-BC3D-9EB70F8356C1}"
 EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ZeldaTest", "ZeldaTest\ZeldaTest.vcxproj", "{D5580B30-FA74-4F0D-ABAA-90CB33EFD703}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|x64 = Debug|x64
@@ -21,6 +23,14 @@ Global
 		{EBC6D2CA-29E7-4666-BC3D-9EB70F8356C1}.Release|x64.Build.0 = Release|x64
 		{EBC6D2CA-29E7-4666-BC3D-9EB70F8356C1}.Release|x86.ActiveCfg = Release|Win32
 		{EBC6D2CA-29E7-4666-BC3D-9EB70F8356C1}.Release|x86.Build.0 = Release|Win32
+		{D5580B30-FA74-4F0D-ABAA-90CB33EFD703}.Debug|x64.ActiveCfg = Debug|x64
+		{D5580B30-FA74-4F0D-ABAA-90CB33EFD703}.Debug|x64.Build.0 = Debug|x64
+		{D5580B30-FA74-4F0D-ABAA-90CB33EFD703}.Debug|x86.ActiveCfg = Debug|Win32
+		{D5580B30-FA74-4F0D-ABAA-90CB33EFD703}.Debug|x86.Build.0 = Debug|Win32
+		{D5580B30-FA74-4F0D-ABAA-90CB33EFD703}.Release|x64.ActiveCfg = Release|x64
+		{D5580B30-FA74-4F0D-ABAA-90CB33EFD703}.Release|x64.Build.0 = Release|x64
+		{D5580B30-FA74-4F0D-ABAA-90CB33EFD703}.Release|x86.ActiveCfg = Release|Win32
+		{D5580B30-FA74-4F0D-ABAA-90CB33EFD703}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
