+++ ZeldaTest/test.cpp
@@ -1,0 +1,37 @@
+#include "pch.h"
+#include "../Zelda/Zelda.cpp"
+
+TEST(TestCaseName, TestName) {
+  EXPECT_EQ(1, 1);
+  EXPECT_TRUE(true);
+}
+
+//---------------------Castle------------------------
+
+TEST(CastleLinking, LinkRoom5and8) {
+	Castle* castle = new Castle();
+
+	// paths[0] = North
+	// paths[1] = South
+	// paths[2] = East
+	// paths[3] = West
+
+	// Check that the rooms are not currently linked
+	Room* room5Paths = castle->getRoom(5)->getPaths()[1];
+	Room* room8Paths = castle->getRoom(8)->getPaths()[0];
+
+	EXPECT_EQ(room5Paths, nullptr);
+	EXPECT_EQ(room8Paths, nullptr);
+
+	// Link the rooms
+	castle->LinkRoom5and8();
+
+	// Test that they were successfully linked
+	room5Paths = castle->getRoom(5)->getPaths()[1];
+	room8Paths = castle->getRoom(8)->getPaths()[0];
+
+	EXPECT_EQ(room5Paths, castle->getRoom(8));
+	EXPECT_EQ(room8Paths, castle->getRoom(5));
+
+	delete castle;
+}
